pyproject.toml
[tool.poetry]
name = "reporttask"
version = "0.1.0"
description = "To manage reports asynchronously"
authors = ["Vibhor Jain"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
xhtml2pdf = "^0.2.16"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
celery = {extras = ["redis"], version = "^5.4.0"}
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
